requirements.txt
annotated-types==0.6.0
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.11
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.1.7
django-ajax-datatable==4.4.5
django-bootstrap-modal-forms==3.0.4
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-extensions==3.2.1
django-filter==23.5
django-htmx==1.14.0
django-import-export==3.3.3
django-js-asset==2.1.0
django-mptt==0.15.0
django-ninja==1.1.0
django-rest-framework==0.1.0
django-tables2==2.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.2.2
djlint==1.34.1
EditorConfig==0.12.3
et-xmlfile==1.1.0
filelock==3.12.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.6
jsbeautifier==1.14.11
json5==0.9.14
MarkupPy==1.14
msgpack==1.0.5
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.1
pathspec==0.12.1
Pillow==10.1.0
pipenv==2023.4.29
platformdirs==3.5.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.7.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.1
redis==4.5.5
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tablib==3.5.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2022.7
urllib3==2.1.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
